ADDED ROUTER FOR GAME LIST PAGE
Created a controller file which handles the routing for the game list page. Fixed an invalid syntax issue in the games table. Assigned the port to a variable to avoid linter error. Added the list controller to the app.js file.

The list controller returns an annonymous javascript array of objects which may be changed later on.