Add static routing middleware
Needed to create a custom middleware because koa-static doesn't let
you customise the base URL for serving static files. This means
that it has to live at /, when we would ideally like home to live
there.

Added the serve middleware in ./controllers/middleware/serve.js and
a mimetype detection function in ./utils/mime.js