Fix issue with serving static images
Modified controllers/middleware/serve.js because it didn't serve image
files correctly. This was due to telling the fs.readFileSync to open all
files with 'utf8' encoding which would only apply to text files. The
issue was resolved by removing utf8 default encoding and setting the
response body to the raw buffer returned by fs.readFileSync.
Also added some integration tests to verify that the correct mimetype
and contents is being served by the middleware when used in a Koa app,
as well as testing configuration settings for the middleware.