Script to add games by running 'npm run add-data'
Added a new add-data.js file to the /build directory. Gets a list of steam games and adds them to the database one by one. Updated some routes and hbs files so that images can also be shows as URLS instead of paths to /images. Currently adds only 100 games, but can be increased by changing the 'gameAmount` variable. There's a restriction on how many requests you can make to the steam api.